refactor: LandingHeader 변경
