+++ src/components/layout/LandingHeader/index.tsx
@@ -1,19 +1,21 @@
 'use client';
 
-import Link from 'next/link';
-import LogoIcon from '@/public/images/icon-logo.svg';
+import LogoIcon from '@/public/images/logo.svg';
 import * as S from './styles';
 
 export default function LandingHeader() {
   return (
-    <S.Wrapper>
-      <Link href="/">
-        <LogoIcon />
-      </Link>
-      <S.NavLinks>
-        <Link href="/login">로그인</Link>
-        <Link href="/signup">회원가입</Link>
-      </S.NavLinks>
-    </S.Wrapper>
+    <S.Header>
+      <S.HdInner>
+        <S.LogoBox>
+          <LogoIcon />
+        </S.LogoBox>
+
+        <S.Util>
+          <S.UtilLink href="/login">로그인</S.UtilLink>
+          <S.UtilLink href="/signup">회원가입</S.UtilLink>
+        </S.Util>
+      </S.HdInner>
+    </S.Header>
   );
 }
