+++ src/components/layout/LandingHeader/styles.ts
@@ -1,23 +1,56 @@
+import Link from 'next/link';
 import styled from 'styled-components';
+import { DEVICE } from '@/src/styles/Breakpoints';
 
-export const Wrapper = styled.header`
-  position: sticky;
+export const Header = styled.header`
+  z-index: 999;
+  position: fixed;
   top: 0;
-  width: 100%;
-  height: 60px;
-  padding: 0 24px;
-  background: var(--color-gray-100);
-  border-bottom: 1px solid var(--color-gray-300);
-  z-index: 100;
+  left: 0;
+  right: 0;
+  height: 96px;
+  background: #fff;
+  border-bottom: 1px solid #1c1b20;
 
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
+  @media ${DEVICE.tablet} {
+    height: 67px;
+  }
 `;
 
-export const NavLinks = styled.nav`
+export const HdInner = styled.div`
+  margin: 0 auto;
+  padding: 0 20px;
+  max-width: 1740px;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+export const LogoBox = styled.div`
   display: flex;
   align-items: center;
-  gap: 16px;
-  font: var(--md-14px-medium);
+  gap: 6px;
+`;
+
+export const LogoImage = styled.img`
+  width: 186px;
+
+  @media ${DEVICE.tablet} {
+    width: 130px;
+  }
+`;
+
+export const Util = styled.nav`
+  display: flex;
+  align-items: center;
+  gap: 14px;
+`;
+
+export const UtilLink = styled(Link)`
+  padding: 10px 12px;
+  font-size: 16px;
+  color: #404040;
+  font-weight: 500;
 `;
